# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobListResponse": {
        "description": "GET /v1/jobs - one page of jobs, newest first.\n\nnext_cursor is opaque: pass it back verbatim as ?cursor= to fetch the\nfollowing page. None means this was the last page.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
      "JobOptimizationResponse": {
        "description": "Honest outcome of one corridor optimization batch.\n\nReports scheduled and unscheduled work separately so a partially\nsuccessful run can never read as a fully successful one.",
        "properties": {
          "corridor_id": {
            "title": "Corridor Id",
            "type": "string"
          },
          "counts": {
            "$ref": "#/components/schemas/OptimizationCounts"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "infeasibility_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Infeasibility Reasons",
            "type": "array"
          },
          "optimization_run_id": {
            "title": "Optimization Run Id",
            "type": "string"
          },
          "possession_derivation": {
            "title": "Possession Derivation",
            "type": "string"
          },
          "possession_source": {
            "title": "Possession Source",
            "type": "string"
          },
          "possession_window_count": {
            "title": "Possession Window Count",
            "type": "integer"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceResponse"
          },
          "scheduled": {
            "items": {
              "$ref": "#/components/schemas/ScheduledJobOutcome"
            },
            "title": "Scheduled",
            "type": "array"
          },
          "solve_time_seconds": {
            "title": "Solve Time Seconds",
            "type": "number"
          },
          "solver_status": {
            "title": "Solver Status",
            "type": "string"
          },
          "uncovered_tracks": {
            "items": {
              "type": "string"
            },
            "title": "Uncovered Tracks",
            "type": "array"
          },
          "unscheduled": {
            "items": {
              "$ref": "#/components/schemas/UnscheduledJobOutcome"
            },
            "title": "Unscheduled",
            "type": "array"
          }
        },
        "required": [
          "corridor_id",
          "optimization_run_id",
          "solver_status",
          "solve_time_seconds",
          "possession_source",
          "possession_derivation",
          "possession_window_count",
          "provenance",
          "uncovered_tracks",
          "generated_at",
          "counts",
          "scheduled",
          "unscheduled",
          "infeasibility_reasons"
        ],
        "title": "JobOptimizationResponse",
        "type": "object"
      },
...
      "JobStatus": {
        "enum": [
          "reported",
          "scheduled",
          "notified",
          "in_progress",
          "completed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "OptimizationCounts": {
        "properties": {
          "committed": {
            "title": "Committed",
            "type": "integer"
          },
          "considered": {
            "title": "Considered",
            "type": "integer"
          },
          "scheduled": {
            "title": "Scheduled",
            "type": "integer"
          },
          "unscheduled": {
            "title": "Unscheduled",
            "type": "integer"
          }
        },
        "required": [
          "considered",
          "committed",
          "scheduled",
          "unscheduled"
        ],
        "title": "OptimizationCounts",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Pashupatastra API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/corridors/{corridor_id}/optimize-jobs": {
      "post": {
        "description": "Run the existing CP-SAT solver over this corridor's active jobs.\n\nReturns an honest batch outcome: scheduled and unscheduled work are\nreported separately, with the solver's own reasons, and the\nprovenance of the possession data used is stated explicitly.",
        "operationId": "optimize_corridor_jobs_v1_corridors__corridor_id__optimize_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "corridor_id",
            "required": true,
            "schema": {
              "title": "Corridor Id",
              "type": "string"
            }
          },
          {
            "description": "Declared actor identifier (e.g. WORKER-042). NOT verified: recorded with DECLARED_UNVERIFIED assurance. Send together with X-Actor-Role, or send neither.",
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Declared actor identifier (e.g. WORKER-042). NOT verified: recorded with DECLARED_UNVERIFIED assurance. Send together with X-Actor-Role, or send neither.",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Declared role: WORKER, ENGINEER, AUTHORITY or ADMIN. SYSTEM is refused (403). NOT verified and not yet enforced as a permission: authentication and RBAC are later slices.",
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Declared role: WORKER, ENGINEER, AUTHORITY or ADMIN. SYSTEM is refused (403). NOT verified and not yet enforced as a permission: authentication and RBAC are later slices.",
              "title": "X-Actor-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOptimizationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Corridor Jobs",
        "tags": [
          "jobs",
          "jobs"
        ]
      }
    },
    "/v1/jobs": {
      "get": {
        "description": "One page of jobs, newest first: {items, next_cursor}.\n\nOrdered by (created_at, job_id) descending, so paging never repeats\nor skips a job that existed when paging began. next_cursor is null on\nthe last page.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "description": "Optional status filter, e.g. 'reported' for the authority review queue.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional status filter, e.g. 'reported' for the authority review queue.",
              "title": "Status"
            }
          },
          {
            "description": "Page size (1-200, default 50).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size (1-200, default 50).",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor: the next_cursor of the previous page, passed back verbatim. Omit for the first page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor: the next_cursor of the previous page, passed back verbatim. Omit for the first page.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "create_job_v1_jobs_post",
        "parameters": [
          {
            "description": "Declared actor identifier (e.g. WORKER-042). NOT verified: recorded with DECLARED_UNVERIFIED assurance. Send together with X-Actor-Role, or send neither.",
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Declared actor identifier (e.g. WORKER-042). NOT verified: recorded with DECLARED_UNVERIFIED assurance. Send together with X-Actor-Role, or send neither.",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Declared role: WORKER, ENGINEER, AUTHORITY or ADMIN. SYSTEM is refused (403). NOT verified and not yet enforced as a permission: authentication and RBAC are later slices.",
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Declared role: WORKER, ENGINEER, AUTHORITY or ADMIN. SYSTEM is refused (403). NOT verified and not yet enforced as a permission: authentication and RBAC are later slices.",
              "title": "X-Actor-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}